requirements/dev.txt
Django==1.9
django-debug-toolbar==1.5
pkg-resources==0.0.0
sqlparse==0.2.0
django-imagekit==3.3
pilkit==1.1.13
six==1.10.0
